Refactor invoke commands - split python code to a separate modules; Add `print-mp3-tags` command
